Coordinates

+++ lib/src/directions.response.dart
@@ -90,8 +90,8 @@ class Coordinates {
 
   factory Coordinates.fromMap(Map<String, dynamic> map) {
     return Coordinates(
-      latitude: (map['latLng']['latitude']).toDouble(),
-      longitude: (map['latLng']['longitude']).toDouble(),
+      latitude: map['latLng']['latitude'],
+      longitude: map['latLng']['longitude'],
     );
   }
 }
